@@ -1,6 +1,6 @@
 {
 	"name": "vue-parcel",
-	"description": "Simple boilerplate for building static SPA's with VueJs 2.",
+	"description": "Simple boilerplate for building static SPA's with VueJs 3.",
 	"license": "UNLICENSED",
 	"main": "app/app.js",
 	"repository": "https://github.com/npulidom/vue-parcel",
@@ -20,19 +20,15 @@
 		"@vue/component-compiler-utils": "~3.3.0",
 		"autoprefixer": "~10.4.2",
 		"cssnano": "~5.0.17",
-		"parcel-config-vue2": "~0.1.3",
-		"parcel-optimizer-css-url": "~2.0.2",
-		"parcel-transformer-vue2": "~0.1.7",
+		"parcel-resolver-vue3-runtime": "~2.0.2",
 		"postcss": "~8.4.6",
 		"pug": "~3.0.2",
-		"stylus": "~0.57.0",
-		"vue-hot-reload-api": "~2.3.4",
-		"vue-template-compiler": "~2.6.14"
+		"stylus": "~0.57.0"
 	},
 	"dependencies": {
 		"modern-normalize": "~1.1.0",
-		"vue": "~2.6.14",
-		"vue-router": "~3.5.3",
-		"vuex": "~3.6.2"
+		"vue": "~3.2.31",
+		"vue-router": "~4.0.12",
+		"vuex": "~4.0.2"
 	}
 }
