@@ -4,37 +4,9 @@
 
 <script>
 
-import Vue       from "vue"
-import VueRouter from "vue-router"
-
-import Home    from "./Home.vue"
-import Section from "./Section.vue"
-
-Vue.use(VueRouter)
-
-// Routes
-const routes = [
-
-	{
-		path     : "/",
-		component:  Home
-	},
-	{
-		path     : "/section",
-		component: Section
-	},
-	// not found
-	{
-		path    : "*",
-		redirect: "/"
-	}
-]
-
 export default {
 
 	name: "App",
-
-	router: new VueRouter({ routes })
 }
 </script>
 
@@ -42,14 +14,17 @@ export default {
 
 #app
 
-	transition(name="fade" mode="out-in" :duration="{ enter: 300, leave: 300 }" appear)
-		router-view
+	//- Router
+	router-view(v-slot="{ Component }")
+		transition(name="fade" mode="out-in")
+			component(:is="Component")
 
 </template>
 
-<style lang="stylus">
+<style lang="stylus" scoped>
 
 #app
+	width 100vw
 	min-height 100vh
 
 </style>
